cleaned up the code, still need to comment
