+++ todolist.txt
@@ -1,2 +1,3 @@
-1:do the laundry
-0:nothing
+0:take car to the shop
+0:add comments to the code
+1:tidy up the code
